+++ src/raycasting/draw_minimap.c
@@ -6,7 +6,7 @@
 /*   By: avaganay <avaganay@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/14 15:42:27 by avaganay          #+#    #+#             */
-/*   Updated: 2023/06/17 12:04:53 by avaganay         ###   ########.fr       */
+/*   Updated: 2023/06/17 13:40:49 by avaganay         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void	my_mlx_pixel_put(t_data *data, int x, int y, int color)
 	*(unsigned int*)dst = color;
 }
 
-void	draw_playeur(t_all *all, int i, int j)
+void	draw_playeur(t_all *all, double i, double j)
 {
 	int	x;
 	int	y;
