+++ src/raycasting/refresh.c
@@ -6,7 +6,7 @@
 /*   By: avaganay <avaganay@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/17 12:00:30 by avaganay          #+#    #+#             */
-/*   Updated: 2023/06/17 12:00:46 by avaganay         ###   ########.fr       */
+/*   Updated: 2023/06/17 13:40:34 by avaganay         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,6 @@ void	refresh(t_all *all)
     mlx_clear_window(all->mlx,all->mlx_win);
 	mlx_put_image_to_window(all->mlx, all->mlx_win, all->data->img, 0, 0);
 	draw_map(all);
-	draw_playeur(all,(int)all->ray->pos.x, (int)all->ray->pos.y);
+	draw_playeur(all,all->ray->pos.x, all->ray->pos.y);
 	raycasting(all);
 }
